+++ CHANGELOG.md
@@ -2,7 +2,19 @@
 
 ## [Unreleased][]
 
-[Unreleased]: https://github.com/chaostoolkit-incubator/chaostoolkit-google-cloud-platform/compare/0.9.2...HEAD
+[Unreleased]: https://github.com/chaostoolkit-incubator/chaostoolkit-google-cloud-platform/compare/0.10.0...HEAD
+
+## [0.10.0][] - 2023-08-22
+
+[0.10.0]: https://github.com/chaostoolkit-incubator/chaostoolkit-google-cloud-platform/compare/0.9.2...0.10.0
+
+### Added
+
+* Added fault injection policiy actions for UrlMap/LB
+
+### Changed
+
+* Switched from flake8 to ruff
 
 ## [0.9.2][] - 2023-05-18